Fix FileController namespace and constructor
