+++ lib/Controller/FileController.php
@@ -1,5 +1,5 @@
 <?php
-namespace OCA\nextcloudscanworkflow\Controller;
+namespace OCA\Nextcloudscanworkflow\Controller;
 
 use OCP\AppFramework\Http\DataResponse;
 use OCP\AppFramework\ApiController;
@@ -8,6 +8,14 @@ use OCP\IRequest;
 use OCP\IUserSession;
 
 class FileController extends ApiController {
+    private $rootFolder;
+    private $userId;
+
+    public function __construct(IRequest $request, IRootFolder $rootFolder, IUserSession $userSession) {
+        parent::__construct('nextcloudscanworkflow', $request);
+        $this->rootFolder = $rootFolder;
+        $this->userId = $userSession->getUser()->getUID();
+    }
 
     /**
      * @NoAdminRequired
