--- lib/controller/FileController.php
@@ -1,38 +1,0 @@
-<?php
-namespace OCA\nextcloudscanWorkflow\Controller;
-
-use OCP\AppFramework\Http\DataResponse;
-use OCP\AppFramework\ApiController;
-use OCP\Files\IRootFolder;
-use OCP\IRequest;
-use OCP\IUserSession;
-
-class FileController extends ApiController {
-    private $rootFolder;
-    private $userId;
-
-    public function __construct(IRequest $request, IRootFolder $rootFolder, IUserSession $userSession) {
-        parent::__construct('pdfworkflow', $request);
-        $this->rootFolder = $rootFolder;
-        $this->userId = $userSession->getUser()->getUID();
-    }
-
-    /**
-     * @NoAdminRequired
-     */
-    public function listFiles(): DataResponse {
-        $folder = $this->rootFolder->getUserFolder($this->userId)->get('inbox');
-        $files = [];
-
-        foreach ($folder->getDirectoryListing() as $file) {
-            if ($file->getMimeType() === 'application/pdf') {
-                $files[] = [
-                    'name' => $file->getName(),
-                    'path' => $file->getPath(),
-                ];
-            }
-        }
-
-        return new DataResponse($files);
-    }
-}
